fixed error

+++ app/src/main/java/org/kiimo/me/main/sender/model/request/pay/UserStatus.kt
@@ -5,5 +5,5 @@ import com.google.gson.annotations.SerializedName
 
 data class UserStatus(
     @SerializedName("avgRating")
-    val avgRating: Int = 0
+    val avgRating: Float = 0f
 )
